chore: Move type tests away from src
There is no need for them to be in src since they test against
the built types in dist.
Moving them out cleans it all up and allows for shorter relative import paths.
